move shared methods to PArrayBackedContainer
